requirements.txt
faiss-cpu==1.8.0.post1
fastapi==0.112.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.99
openai==1.41.0
pypdf==5.0.1
pdfplumber==0.11.4
tiktoken=0.8.0
IPython==8.28.0
python-dotenv==1.0.1
mysql-connector-python==9.1.0